Move updates module to app
